pyproject.toml
[project]
name = "auto-anki"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.81.0",
    "pylint>=3.3.7",
    "python-dotenv>=1.1.0",
]
